Refactor du layout principal : suppression des dépendances inutiles, intégration de ClientLayout et exportation des métadonnées. Simplification du code pour une meilleure lisibilité.
