+++ src/app/metadata.ts
@@ -1,0 +1,40 @@
+import type { Metadata } from "next"
+
+export const metadata: Metadata = {
+  title: "Lucas Spitzer - Étudiant BTS SIO & Développeur Full Stack | Portfolio",
+  description: "Portfolio de Lucas Spitzer, étudiant en BTS SIO SLAM et développeur full stack passionné. Découvrez mes projets web, mes compétences en développement et mon parcours en formation.",
+  keywords: "étudiant BTS SIO, BTS SIO SLAM, développeur web junior, full stack, React, TypeScript, PHP, portfolio étudiant, développeur, Nîmes, formation développement web, alternance développement, stage développement web, projets web étudiants",
+  authors: [{ name: "Lucas Spitzer" }],
+  creator: "Lucas Spitzer",
+  publisher: "Lucas Spitzer",
+  icons: {
+    icon: '/favicon.ico',
+  },
+  formatDetection: {
+    email: false,
+    address: false,
+    telephone: false,
+  },
+  openGraph: {
+    title: "Lucas Spitzer - Étudiant BTS SIO & Développeur Full Stack | Portfolio",
+    description: "Portfolio de Lucas Spitzer, étudiant en BTS SIO SLAM et développeur full stack passionné. Découvrez mes projets web, mes compétences en développement et mon parcours en formation.",
+    type: "website",
+    locale: "fr_FR",
+    siteName: "Portfolio de Lucas Spitzer - Étudiant BTS SIO",
+  },
+  other: {
+    "github-profile": "https://github.com/spitzerl",
+    "linkedin-profile": "https://www.linkedin.com/in/lucasspitzer/",
+  },
+  robots: {
+    index: true,
+    follow: true,
+    googleBot: {
+      index: true,
+      follow: true,
+      "max-video-preview": -1,
+      "max-image-preview": "large",
+      "max-snippet": -1,
+    },
+  },
+};
